+++ framework/logstore/migrations.go
@@ -133,6 +133,9 @@ func triggerMigrations(ctx context.Context, db *gorm.DB) error {
 	if err := migrationAddListModelsOutputColumn(ctx, db); err != nil {
 		return err
 	}
+	if err := migrationAddRoutingEngineLogsColumn(ctx, db); err != nil {
+		return err
+	}
 	return nil
 }
 
@@ -1171,6 +1174,39 @@ func migrationAddListModelsOutputColumn(ctx context.Context, db *gorm.DB) error 
 	err := m.Migrate()
 	if err != nil {
 		return fmt.Errorf("error while adding list models output column: %s", err.Error())
+	}
+	return nil
+}
+
+func migrationAddRoutingEngineLogsColumn(ctx context.Context, db *gorm.DB) error {
+	opts := *migrator.DefaultOptions
+	opts.UseTransaction = true
+	m := migrator.New(db, &opts, []*migrator.Migration{{
+		ID: "logs_add_routing_engine_logs_column",
+		Migrate: func(tx *gorm.DB) error {
+			tx = tx.WithContext(ctx)
+			migrator := tx.Migrator()
+			if !migrator.HasColumn(&Log{}, "routing_engine_logs") {
+				if err := migrator.AddColumn(&Log{}, "routing_engine_logs"); err != nil {
+					return err
+				}
+			}
+			return nil
+		},
+		Rollback: func(tx *gorm.DB) error {
+			tx = tx.WithContext(ctx)
+			migrator := tx.Migrator()
+			if migrator.HasColumn(&Log{}, "routing_engine_logs") {
+				if err := migrator.DropColumn(&Log{}, "routing_engine_logs"); err != nil {
+					return err
+				}
+			}
+			return nil
+		},
+	}})
+	err := m.Migrate()
+	if err != nil {
+		return fmt.Errorf("error while adding routing engine logs column: %s", err.Error())
 	}
 	return nil
 }